refactor(bot-mode): split hosted room contract and storage
